+++ src/main/java/chess/controller/GameRestController.java
@@ -1,6 +1,6 @@
 package chess.controller;
 
-import chess.controller.dto.ReachablePositionsDto;
+import chess.dto.ReachablePositionsDto;
 import chess.domain.board.position.Position;
 import chess.domain.piece.Owner;
 import chess.service.GameService;
