+++ src/main/java/chess/dto/ReachablePositionsDto.java
@@ -1,4 +1,4 @@
-package chess.controller.dto;
+package chess.dto;
 
 import java.util.List;
