+++ src/main/java/chess/service/PlayerService.java
@@ -1,6 +1,6 @@
 package chess.service;
 
-import chess.controller.dto.RoomDto;
+import chess.dto.RoomDto;
 import chess.domain.piece.Owner;
 import chess.service.dao.RoomDao;
 import org.springframework.stereotype.Component;
@@ -19,28 +19,47 @@ public class PlayerService {
 
     public void enter(final long roomId, final String playerId) {
         final List<String> players = roomDao.players(roomId);
-        if (players.contains(playerId)) {
+
+        if (isAlreadyJoined(players, playerId)) {
             return;
         }
-        if (players.size() == MAXIMUM_SIZE_OF_ROOM) {
+
+        if(isFull(players)){
             throw new IllegalArgumentException("이미 가득 찬 방입니다.");
         }
+
         roomDao.enter(roomId, playerId);
     }
 
-    public Owner ownerOfPlayer(final Long roomId, final String playerId) {
-        final RoomDto roomInfo = roomDao.roomInfo(roomId);
-        final String player1 = roomInfo.getPlayer1();
-        final String player2 = roomInfo.getPlayer2();
+    private boolean isAlreadyJoined(final List<String> players, final String playerId) {
+        return players.contains(playerId);
+    }
 
-        if (player1.equals(playerId)) {
+    private boolean isFull(final List<String> players){
+        return players.size() == MAXIMUM_SIZE_OF_ROOM;
+    }
+
+    public Owner ownerOfPlayer(final long roomId, final String playerId) {
+        final RoomDto roomInfo = roomDao.roomInfo(roomId);
+
+        if (isPlayer1(roomInfo, playerId)) {
             return Owner.WHITE;
         }
 
-        if (player2.equals(playerId)) {
+        if (isPlayer2(roomInfo, playerId)) {
             return Owner.BLACK;
         }
 
         throw new IllegalArgumentException("적절하지 않은 사용자입니다.");
+    }
+
+    private boolean isPlayer1(final RoomDto roomInfo, final String playerId){
+        final String player1 = roomInfo.getPlayer1();
+        return player1.equals(playerId);
+    }
+
+    private boolean isPlayer2(final RoomDto roomInfo, final String playerId){
+        final String player2 = roomInfo.getPlayer2();
+        return player2.equals(playerId);
     }
 }
