+++ src/test/java/chess/controller/RoomRestControllerTest.java
@@ -1,8 +1,7 @@
 package chess.controller;
 
-import chess.controller.dto.RoomDto;
+import chess.dto.RoomDto;
 import chess.service.RoomService;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
